requirements.txt
flask>=3.0.0
pyyaml>=6.0.1
cryptography>=42.0.0
rich>=13.7.0